moves token simualtion action into redux slice, adds send settings simulation rejection error
