+++ extension/src/popup/App.tsx
@@ -8,6 +8,7 @@ import { metricsMiddleware } from "helpers/metrics";
 import { reducer as auth } from "popup/ducks/accountServices";
 import { reducer as settings } from "popup/ducks/settings";
 import { reducer as transactionSubmission } from "popup/ducks/transactionSubmission";
+import { reducer as tokenPaymentSimulation } from "popup/ducks/token-payment";
 
 import { Loading } from "popup/components/Loading";
 import { ErrorTracking } from "popup/components/ErrorTracking";
@@ -20,6 +21,7 @@ const rootReducer = combineReducers({
   auth,
   settings,
   transactionSubmission,
+  tokenPaymentSimulation,
 });
 export type AppState = ReturnType<typeof rootReducer>;
 export const store = configureStore({
