+++ extension/src/popup/components/sendPayment/SendSettings/SettingsFail/styles.scss
@@ -1,0 +1,112 @@
+.SettingsFail {
+  &__content {
+    flex-grow: 1;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+  }
+
+  &__header {
+    font-size: 0.875rem;
+    line-height: 1.375rem;
+    margin-bottom: 0.25rem;
+  }
+
+  &__amount {
+    font-size: 1.5rem;
+    line-height: 2.25rem;
+    color: var(--color-gray-90);
+    margin-bottom: 1.5rem;
+    text-align: center;
+  }
+
+  &__icon {
+    margin-bottom: 1.5rem;
+
+    svg {
+      height: 40px;
+      width: 40px;
+      stroke-width: 1.5px;
+    }
+  }
+
+  &__success {
+    svg {
+      fill: var(--color-green-60);
+    }
+  }
+
+  &__fail {
+    svg {
+      fill: var(--color-red-60);
+    }
+  }
+
+  &__identicon {
+    display: flex;
+    align-items: center;
+    border: 1px solid var(--color-gray-10);
+    border-radius: 100px;
+    padding: 0.25rem 0.5rem;
+
+    %identiconImg {
+      border-radius: 1rem;
+      width: 0.875rem;
+      height: 0.875rem;
+      margin-right: 0.5rem;
+    }
+
+    span {
+      color: var(--color-gray-90);
+      font-size: 0.875rem;
+      line-height: 1.375rem;
+    }
+  }
+
+  &__suggest-remove-tl {
+    .remove-tl-contents {
+      display: flex;
+      flex-direction: column;
+
+      p {
+        font-size: 14px;
+        margin: 0;
+        line-height: 1.375rem;
+      }
+
+      button {
+        font-size: 14px;
+        font-weight: bold;
+        border: 0;
+        padding: 0;
+        text-align: initial;
+        margin-top: 10px;
+        background-color: transparent;
+        text-decoration: underline;
+        cursor: pointer;
+      }
+    }
+  }
+
+  &__button-rows {
+    &__success {
+      display: flex;
+      column-gap: 1rem;
+      button {
+        flex: 1;
+      }
+    }
+  }
+
+  &__error-code {
+    font-size: 0.75rem;
+    line-height: 1.25rem;
+  }
+
+  &__error-block {
+    margin-top: auto;
+    font-size: 0.875rem;
+  }
+}
